move DisplayBookmarks.js to correct location
